+++ packages/plugin-core/src/dendronExtensionInterface.ts
@@ -3,7 +3,7 @@ import {
   WorkspaceSettings,
   WorkspaceType,
 } from "@dendronhq/common-all";
-import { IWorkspaceService } from "@dendronhq/engine-server";
+import { execa, IWorkspaceService } from "@dendronhq/engine-server";
 import vscode from "vscode";
 import { ILookupControllerV3Factory } from "./components/lookup/LookupControllerV3Interface";
 import {
@@ -11,6 +11,7 @@ import {
   ISchemaLookupProviderFactory,
 } from "./components/lookup/LookupProviderV3Interface";
 import { FileWatcher } from "./fileWatcher";
+import { EngineAPIService } from "./services/EngineAPIService";
 import { IEngineAPIService } from "./services/EngineAPIServiceInterface";
 import { NoteTraitService } from "./services/NoteTraitService";
 import { ISchemaSyncService } from "./services/SchemaSyncServiceInterface";
@@ -54,6 +55,10 @@ export type DendronWorkspaceSettings = Partial<{
  * */
 export interface IDendronExtension {
   port?: number;
+  /**
+   * If set, the server sub process that spawned the engine
+   */
+  serverProcess?: execa.ExecaChildProcess<string>;
   context: vscode.ExtensionContext;
   serverWatcher?: vscode.FileSystemWatcher;
   fileWatcher?: FileWatcher;
@@ -127,4 +132,5 @@ export interface IDendronExtension {
    * about registered Note Traits
    */
   get traitRegistrar(): NoteTraitService;
+  setEngine(svc: EngineAPIService): void;
 }
