+++ packages/plugin-core/src/test/MockDendronExtension.ts
@@ -19,6 +19,7 @@ import {
 } from "../components/lookup/LookupProviderV3Interface";
 import { IDendronExtension } from "../dendronExtensionInterface";
 import { FileWatcher } from "../fileWatcher";
+import { EngineAPIService } from "../services/EngineAPIService";
 import { IEngineAPIService } from "../services/EngineAPIServiceInterface";
 import { NoteTraitService } from "../services/NoteTraitService";
 import { ISchemaSyncService } from "../services/SchemaSyncServiceInterface";
@@ -53,6 +54,10 @@ export class MockDendronExtension implements IDendronExtension {
     this._vaults = vaults;
   }
   get traitRegistrar(): NoteTraitService {
+    throw new Error("Method not implemented.");
+  }
+  serverProcess?: undefined;
+  setEngine(_svc: EngineAPIService): void {
     throw new Error("Method not implemented.");
   }
   fileWatcher?: FileWatcher | undefined;
